requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2024.3.10
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
pexpect==4.9.0
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.6

# OPCIONAL (para produção com Celery + Redis):
# celery==5.3.4
# redis==5.0.1
# flower==2.0.1

# OPCIONAL (para logging avançado):
# python-json-logger==2.0.7
# loguru==0.7.2